# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API справочника организаций",
        "description": "Документация REST API справочника организаций, зданий и деятельности",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/buildings": {
            "get": {
                "tags": [
                    "Buildings"
                ],
                "summary": "Получить список зданий",
                "operationId": "704d1f7897ed7a3b8d29dacea66823ad",
                "responses": {
                    "200": {
                        "description": "Список зданий успешно получен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Building"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Buildings"
                ],
                "summary": "Создать новое здание",
                "operationId": "c8f8399919e01b1e4e2e2f6a9f1f5107",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "address",
                                    "latitude",
                                    "longitude"
                                ],
                                "properties": {
                                    "address": {
                                        "type": "string",
                                        "example": "г. Москва, ул. Ленина 1, офис 3"
                                    },
                                    "latitude": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 55.751244
                                    },
                                    "longitude": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 37.618423
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Здание успешно создано",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Building"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/buildings/{id}": {
            "get": {
                "tags": [
                    "Buildings"
                ],
                "summary": "Получить информацию о здании",
                "operationId": "7c2e9a5ec44b127b681717d1d5b91152",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Информация о здании",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Building"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Здание не найдено",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Здание не найдено"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Buildings"
                ],
                "summary": "Обновить данные здания",
                "operationId": "0bd50384a833589efa669b51b8e1ff88",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BuildingUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Обновленные данные здания",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Building"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Здание не найдено",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Здание не найдено"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Buildings"
                ],
                "summary": "Удалить здание",
                "operationId": "db5360c892964cf865373a95f5da4633",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Здание успешно удалено",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Здание успешно удалено"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Здание не найдено",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Здание не найдено"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "BuildingUpdateRequest": {
                "title": "Building Update Request",
                "description": "Данные для обновления здания",
                "properties": {
                    "address": {
                        "type": "string",
                        "example": "г. Москва, ул. Ленина 10"
                    },
                    "latitude": {
                        "type": "number",
                        "format": "float",
                        "example": 55.75222
                    },
                    "longitude": {
                        "type": "number",
                        "format": "float",
                        "example": 37.61556
                    }
                },
                "type": "object"
            },
...
            "Building": {
                "title": "Building",
                "description": "Модель здания",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "address": {
                        "type": "string",
                        "example": "г. Москва, ул. Ленина 1, офис 3"
                    },
                    "latitude": {
                        "type": "number",
                        "format": "float",
                        "example": 55.751244
                    },
                    "longitude": {
                        "type": "number",
                        "format": "float",
                        "example": 37.618423
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-02-05T12:34:56Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-02-05T12:34:56Z"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}